backend/requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
whitenoise==3.3.1
